+++ src/command_modules/azure-cli-redis/HISTORY.rst
@@ -3,13 +3,14 @@
 Release History
 ===============
 
-0.3.3
+0.4.0
 ++++++
 * Added commands for managing firewall-rules (create, update, delete, show, list)
 * Added commands for managing server-link (create, delete, show, list)
 * Added commands for managing patch-schedule (create, update, delete, show)
 * `az redis create` : Support for Availability Zones and Minimum TLS Version
-* Removed 'az redis update-settings' command
+* BREAKING CHANGE: Removed 'az redis update-settings' and 'az redis list-all' command
+* BREAKING CHANGE: Parameter for redis create: 'tenant settings' is not accepted in key[=value] format
 * Added warning message for deprecating 'az redis import-method' command.
 
 0.3.2